+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- try to fix mobile app link for button in checkout page
+
 ## [0.0.4] - 2025-09-03
 
 ### Changed